# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: ZEINA Hyperviseur API
  version: "1.0.0"
  description: |
    REST + WebSocket API du superviseur IoT ZEINA.

    ## Authentification
    1. POST `/v1/auth/login` retourne un access token (15 min) + un refresh token (cookie httpOnly, 7 j).
    2. Les routes protégées attendent l'access token dans `Authorization: Bearer <token>`.
    3. Quand le token expire, POST `/v1/auth/refresh` produit un nouvel access (utilise le cookie).

    ## Modèle d'autorisation
    - **Superadmin** (`is_superadmin=true`) : bypass complet.
    - **Owner du tenant** (`tenant_role=owner`) : accès implicite à tous les sites + gestion users/rôles.
    - **Membre simple** : accès aux sites où il est listé dans `site_members`, selon le rôle.
    - Chaque rôle attribue un niveau (`none`/`read`/`write`) sur 4 fonctionnalités : `dashboard`, `devices`, `rules`, `members`.

    ## Test rapide
    1. Clique « Authorize » en haut, paste l'access token (sans le préfixe Bearer).
    2. Tous les calls authentifiés vont passer.
servers:
  - url: /
    description: |
      Origine courante (relatif). Quand tu importes ce spec dans Bruno/Postman,
      remplace par `http://localhost:5173` (via nginx) ou `http://localhost:3000`
      (API directe) — sans slash final.
...
security:
  - bearerAuth: []
paths:
...
  /v1/sites/{id}/dashboards:
    parameters: [{ $ref: "#/components/parameters/SiteId" }]
    get:
      tags: [Dashboards & widgets]
      summary: Liste des dashboards d'un site
      responses:
        "200": { description: OK, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/DashboardListItem" } } } } }
    post:
      tags: [Dashboards & widgets]
      summary: Créer un dashboard
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateDashboardReq" }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/Dashboard" } } } }
  /v1/dashboards/{id}:
    parameters: [{ $ref: "#/components/parameters/DashboardId" }]
    get:
      tags: [Dashboards & widgets]
      summary: Détail d'un dashboard avec ses widgets
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/DashboardDetail" } } } }
    put:
      tags: [Dashboards & widgets]
      summary: Renommer / décrire un dashboard
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateDashboardReq" }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Dashboard" } } } }
    delete:
      tags: [Dashboards & widgets]
      summary: Supprimer un dashboard et ses widgets
      responses:
        "204": { description: No content }
  /v1/dashboards/{id}/widgets:
    parameters: [{ $ref: "#/components/parameters/DashboardId" }]
    post:
      tags: [Dashboards & widgets]
      summary: Ajouter un widget
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateWidgetReq" }
            examples:
              value:
                value:
                  type: value
                  title: "Humidité — Capteur ambiance"
                  config:
                    device_id: "<uuid>"
                    measurement: humidity
                    unit: "%"
                    decimals: 1
              gauge:
                value:
                  type: gauge
                  title: "CO₂"
                  config: { device_id: "<uuid>", measurement: co2, unit: ppm, min: 400, max: 2000 }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/Widget" } } } }
  /v1/widgets/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string, format: uuid }
    put:
      tags: [Dashboards & widgets]
      summary: Mettre à jour un widget
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateWidgetReq" }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Widget" } } } }
    delete:
      tags: [Dashboards & widgets]
      summary: Supprimer un widget
      responses:
        "204": { description: No content }
  /v1/dashboards/{id}/layouts:
    parameters: [{ $ref: "#/components/parameters/DashboardId" }]
    put:
      tags: [Dashboards & widgets]
      summary: Sauvegarder la layout (positions x/y/w/h des widgets)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateLayoutsReq" }
      responses:
        "204": { description: No content }

  # ---------------------------------------------------------------------------
  # RULES
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    SiteId:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
      description: ID du site
...
    DashboardId:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
      description: ID du dashboard
...
  schemas:
...
    Dashboard:
      type: object
      properties:
        id:          { type: string, format: uuid }
        site_id:     { type: string, format: uuid }
        name:        { type: string }
        description: { type: string, nullable: true }
        created_at:  { type: string, format: date-time }
        updated_at:  { type: string, format: date-time }
    DashboardListItem:
      allOf:
        - $ref: "#/components/schemas/Dashboard"
        - type: object
          properties:
            widget_count: { type: integer }
    Widget:
      type: object
      properties:
        id:         { type: string, format: uuid }
        type:       { type: string, enum: [value, line, area, bar, gauge, state] }
        title:      { type: string }
        position:   { type: integer }
        config:     { type: object, additionalProperties: true }
        layout:     { type: object, additionalProperties: true }
        created_at: { type: string, format: date-time }
    DashboardDetail:
      allOf:
        - $ref: "#/components/schemas/Dashboard"
        - type: object
          properties:
            widgets:
              type: array
              items: { $ref: "#/components/schemas/Widget" }
    CreateDashboardReq:
      type: object
      required: [name]
      properties:
        name:        { type: string }
        description: { type: string }
    UpdateDashboardReq:
      type: object
      properties:
        name:        { type: string, nullable: true }
        description: { type: string, nullable: true }
    CreateWidgetReq:
      type: object
      required: [type, title]
      properties:
        type:   { type: string, enum: [value, line, area, bar, gauge, state] }
        title:  { type: string }
        config: { type: object, additionalProperties: true }
    UpdateWidgetReq:
      type: object
      properties:
        title:    { type: string, nullable: true }
        position: { type: integer, nullable: true }
        config:   { type: object, additionalProperties: true }
        layout:   { type: object, additionalProperties: true }
    LayoutEntry:
      type: object
      required: [widget_id, x, y, w, h]
      properties:
        widget_id: { type: string, format: uuid }
        x:         { type: integer }
        y:         { type: integer }
        w:         { type: integer }
        h:         { type: integer }
    UpdateLayoutsReq:
      type: object
      required: [layouts]
      properties:
        layouts:
          type: array
          items: { $ref: "#/components/schemas/LayoutEntry" }

    # --- Rules --------------------------------------------------------------
...